Add `mojaveIntegration.FetchClientInterface`
